requirements.txt
streamlit
pandas
scikit-learn
pypdf
plotly
python-docx
matplotlib